Revert "[FEAT] Pet NFT Animation (#6525)"
This reverts commit 0ef2cd90a6440864886ee4bed92dafdc5279f9ab.
