+++ src/features/world/scenes/BaseScene.ts
@@ -4,8 +4,6 @@ import VirtualJoystick from "phaser3-rex-plugins/plugins/virtualjoystick.js";
 
 import { SQUARE_WIDTH } from "features/game/lib/constants";
 import { BumpkinContainer } from "../containers/BumpkinContainer";
-import { PetContainer } from "../containers/PetContainer";
-import { PetNFTType } from "features/game/types/pets";
 import { interactableModalManager } from "../ui/InteractableModals";
 import { NPCName, NPC_WEARABLES } from "lib/npcs";
 import { npcModalManager } from "../ui/NPCModals";
@@ -110,8 +108,6 @@ export abstract class BaseScene extends Phaser.Scene {
   packetSentAt = 0;
 
   playerEntities: { [sessionId: string]: BumpkinContainer } = {};
-
-  pets: { [sessionId: string]: PetContainer } = {};
 
   colliders?: Phaser.GameObjects.Group;
   triggerColliders?: Phaser.GameObjects.Group;
@@ -969,7 +965,6 @@ export abstract class BaseScene extends Phaser.Scene {
     this.updateShaders();
     this.updateUsernames();
     this.updateFactions();
-    this.updatePets();
   }
 
   keysToAngle(
@@ -1219,73 +1214,6 @@ export abstract class BaseScene extends Phaser.Scene {
           : "#fff";
 
         this.checkAndUpdateNameColor(this.playerEntities[sessionId], color);
-      }
-    });
-  }
-
-  public addPet(
-    sessionId: string,
-    petId: number,
-    petType: string,
-    x: number,
-    y: number,
-  ) {
-    const petContainer = new PetContainer(
-      this,
-      x,
-      y,
-      petId,
-      petType as PetNFTType,
-    );
-    this.pets[sessionId] = petContainer;
-  }
-
-  public updatePets() {
-    const server = this.mmoServer;
-    if (!server) return;
-
-    Object.keys(this.pets).forEach((sessionId) => {
-      const petsMap = server.state.pets;
-      if (!petsMap) return;
-
-      const hasLeft =
-        !petsMap.get(sessionId) ||
-        petsMap.get(sessionId)?.sceneId !== this.scene.key;
-
-      const isInactive = !this.pets[sessionId]?.active;
-
-      if (hasLeft || isInactive) {
-        this.pets[sessionId]?.destroy();
-        delete this.pets[sessionId];
-      }
-    });
-
-    server.state.pets?.forEach((pet, sessionId) => {
-      if (pet.sceneId !== this.scene.key) return;
-
-      const petContainer = this.pets[sessionId];
-      if (!petContainer) {
-        this.addPet(sessionId, pet.id, pet.type, pet.x, pet.y);
-        return;
-      }
-
-      if (petContainer) {
-        const distance = Phaser.Math.Distance.BetweenPoints(petContainer, pet);
-
-        if (distance > 1) {
-          if ((pet.x || 0) > petContainer.x) {
-            petContainer.faceRight();
-          } else if ((pet.x || 0) < petContainer.x) {
-            petContainer.faceLeft();
-          }
-          petContainer.walk();
-        } else {
-          petContainer.idle();
-        }
-
-        petContainer.x = Phaser.Math.Linear(petContainer.x, pet.x, 0.04);
-        petContainer.y = Phaser.Math.Linear(petContainer.y, pet.y, 0.04);
-        petContainer.setDepth(petContainer.y);
       }
     });
   }
